+++ tests/antares/integration/test_local_client.py
@@ -16,6 +16,7 @@ from pathlib import Path
 import numpy as np
 import pandas as pd
 
+from antares.craft import create_study_local
 from antares.craft.exceptions.exceptions import AreaCreationError, LinkCreationError
 from antares.craft.model.area import AdequacyPatchMode, Area, AreaProperties, AreaUi
 from antares.craft.model.binding_constraint import BindingConstraintProperties, ClusterData, ConstraintTerm, LinkData
@@ -29,7 +30,7 @@ from antares.craft.model.settings.advanced_parameters import (
 from antares.craft.model.settings.general import GeneralParametersUpdate
 from antares.craft.model.settings.study_settings import StudySettingsUpdate
 from antares.craft.model.st_storage import STStorageGroup, STStorageProperties
-from antares.craft.model.study import Study, create_study_local
+from antares.craft.model.study import Study
 from antares.craft.model.thermal import ThermalCluster, ThermalClusterGroup, ThermalClusterProperties
 from antares.craft.tools.ini_tool import IniFile, InitializationFilesTypes
