+++ tests/antares/services/api_services/test_binding_constraint_api.py
@@ -20,6 +20,7 @@ from antares.craft.exceptions.exceptions import ConstraintMatrixDownloadError, C
 from antares.craft.model.area import Area
 from antares.craft.model.binding_constraint import BindingConstraint, BindingConstraintProperties, ConstraintMatrixName
 from antares.craft.model.study import Study
+from antares.craft.service.api_services.factory import ApiServiceFactory
 from antares.craft.service.service_factory import ServiceFactory
 
 
@@ -36,14 +37,15 @@ def constraint_set():
 class TestCreateAPI:
     api = APIconf("https://antares.com", "token", verify=False)
     study_id = "22c52f44-4c2a-407b-862b-490887f93dd8"
-    study = Study("study_test", "870", ServiceFactory(api, study_id))
+    factory = ApiServiceFactory(api, study_id)
+    study = Study("study_test", "870", factory)
     area = Area(
         "study_test",
-        ServiceFactory(api, study_id).create_area_service(),
-        ServiceFactory(api, study_id).create_st_storage_service(),
-        ServiceFactory(api, study_id).create_thermal_service(),
-        ServiceFactory(api, study_id).create_renewable_service(),
-        ServiceFactory(api, study_id).create_hydro_service(),
+        factory.create_area_service(),
+        factory.create_st_storage_service(),
+        factory.create_thermal_service(),
+        factory.create_renewable_service(),
+        factory.create_hydro_service(),
     )
     antares_web_description_msg = "Mocked Server KO"
     matrix = pd.DataFrame(data=[[0]])
@@ -52,7 +54,7 @@ class TestCreateAPI:
         with requests_mock.Mocker() as mocker:
             properties = BindingConstraintProperties(enabled=False)
             constraint = BindingConstraint(
-                "bc_1", ServiceFactory(self.api, self.study_id).create_binding_constraints_service()
+                "bc_1", ApiServiceFactory(self.api, self.study_id).create_binding_constraints_service()
             )
             url = f"https://antares.com/api/v1/studies/{self.study_id}/bindingconstraints/{constraint.id}"
             mocker.put(url, json={"id": "id", "name": "name", "terms": [], **properties.model_dump()}, status_code=200)
@@ -62,7 +64,7 @@ class TestCreateAPI:
         with requests_mock.Mocker() as mocker:
             properties = BindingConstraintProperties(enabled=False)
             constraint = BindingConstraint(
-                "bc_1", ServiceFactory(self.api, self.study_id).create_binding_constraints_service()
+                "bc_1", ApiServiceFactory(self.api, self.study_id).create_binding_constraints_service()
             )
             url = f"https://antares.com/api/v1/studies/{self.study_id}/bindingconstraints/{constraint.id}"
             antares_web_description_msg = "Server KO"
@@ -76,7 +78,7 @@ class TestCreateAPI:
 
     def test_get_constraint_matrix_success(self, constraint_set):
         constraint = BindingConstraint(
-            "bc_test", ServiceFactory(self.api, self.study_id).create_binding_constraints_service()
+            "bc_test", ApiServiceFactory(self.api, self.study_id).create_binding_constraints_service()
         )
         for matrix_method, enum_value, path, expected_matrix in constraint_set:
             with requests_mock.Mocker() as mocker:
@@ -87,7 +89,7 @@ class TestCreateAPI:
 
     def test_get_constraint_matrix_fails(self, constraint_set):
         constraint = BindingConstraint(
-            "bc_test", ServiceFactory(self.api, self.study_id).create_binding_constraints_service()
+            "bc_test", ApiServiceFactory(self.api, self.study_id).create_binding_constraints_service()
         )
         for matrix_method, enum_value, path, _ in constraint_set:
             with requests_mock.Mocker() as mocker:
