+++ tests/antares/services/api_services/test_thermal_api.py
@@ -9,12 +9,11 @@
 # SPDX-License-Identifier: MPL-2.0
 #
 # This file is part of the Antares project.
-import pytest
-import requests_mock
-
 from unittest.mock import Mock
 
 import pandas as pd
+import pytest
+import requests_mock
 
 from antares.craft.api_conf.api_conf import APIconf
 from antares.craft.exceptions.exceptions import (
@@ -31,9 +30,9 @@ from antares.craft.model.thermal import (
     ThermalClusterPropertiesUpdate,
 )
 from antares.craft.service.api_services.area_api import AreaApiService
+from antares.craft.service.api_services.factory import ApiServiceFactory
 from antares.craft.service.api_services.models.thermal import ThermalClusterPropertiesAPI
 from antares.craft.service.api_services.services.thermal import ThermalApiService
-from antares.craft.service.service_factory import ServiceFactory
 
 
 @pytest.fixture
@@ -51,16 +50,17 @@ def thermal_matrix_set():
 class TestCreateAPI:
     api = APIconf("https://antares.com", "token", verify=False)
     study_id = "22c52f44-4c2a-407b-862b-490887f93dd8"
-    study = Study("study_test", "870", ServiceFactory(api, study_id))
+    factory = ApiServiceFactory(api, study_id)
+    study = Study("study_test", "870", factory)
     area = Area(
         "area-test",
-        ServiceFactory(api, study_id).create_area_service(),
-        ServiceFactory(api, study_id).create_st_storage_service(),
-        ServiceFactory(api, study_id).create_thermal_service(),
-        ServiceFactory(api, study_id).create_renewable_service(),
-        ServiceFactory(api, study_id).create_hydro_service(),
+        factory.create_area_service(),
+        factory.create_st_storage_service(),
+        factory.create_thermal_service(),
+        factory.create_renewable_service(),
+        factory.create_hydro_service(),
     )
-    thermal = ThermalCluster(ServiceFactory(api, study_id).create_thermal_service(), "area-test", "thermal-test")
+    thermal = ThermalCluster(factory.create_thermal_service(), "area-test", "thermal-test")
     antares_web_description_msg = "Mocked Server KO"
     matrix = pd.DataFrame(data=[[0]])
