+++ tests/antares/services/local_services/test_study_read.py
@@ -16,7 +16,7 @@ import re
 
 from pathlib import Path
 
-from antares.craft.model.study import read_study_local
+from antares.craft import read_study_local
 
 
 class TestReadStudy:
